added alpha to xresources

+++ Documents/Learn/C/Suckless/st-0.8.4/config.h
@@ -162,7 +162,7 @@ ResourcePref resources[] = {
 		{ "borderpx",     INTEGER, &borderpx },
 		{ "cwscale",      FLOAT,   &cwscale },
 		{ "chscale",      FLOAT,   &chscale },
-		{ "alpha",        INTEGER, &alpha },
+		{ "alpha",        FLOAT,   &alpha },
 };
